main-service-master/requirements.txt
aiobotocore>=0.10.3
aiofiles>=0.4.0
aioredis>=1.3.1
alembic>=1.0.11
arrow>=0.14.3
asyncio>=3.4.3
asyncpg>=0.18.3
bcrypt>=3.1.7
Cython>=0.29.13
datadog>=0.29.3
elasticsearch>=7.1.0
email-validator>=1.0.4
fastapi==0.42.0
hiredis>=1.0.0
httpx>=0.7.2
phonenumbers>=8.10.21
prometheus-client>=0.7.1
psycopg2>=2.8.3 
pytest>=5.0.1
pytest-asyncio>=0.10.0
pytest-cov>=2.7.1
pytest-mock>=1.10.4
python-multipart>=0.0.5
python-rapidjson>=0.7.2
requests>=2.22.0
structlog>=19.1.0
starlette-prometheus>=0.5.0 
timing-asgi>=0.2.0
twilio>=6.37.0
uvicorn>=0.8.4
cerberus
requests_pkcs12
xmltodict
